+++ tests-system/lobster-html-report/lobster_UI_test_runner.py
@@ -12,7 +12,6 @@ from webdriver_manager.chrome import ChromeDriverManager
 class CmdArgs:
     lobster_report: Optional[str] = None
     out: Optional[str] = None
-    dot: Optional[str] = None
     high_contrast: Optional[str] = None
     render_md: bool = False
 
@@ -31,7 +30,6 @@ class CmdArgs:
             cmd_args.append("--render-md")
 
         append_if_string("--out", self.out)
-        append_if_string("--dot", self.dot)
         append_if_string("--high-contrast", self.high_contrast)
         return cmd_args